[RER] - (#3) refactoring the file position
